+++ CHANGELOG.md
@@ -32,6 +32,8 @@
   * Secondary navigation removed in favor of terra-application secondary navigation layout.
   * All content now wrapped in a page.
   * Consume terra-cli for releasing.
+  * Removed reference to infinite list
+  * Updated props table to better format types, defaults and description.
 
 ## 6.30.0 - (September 11, 2020)