+++ CHANGELOG.md
@@ -35,6 +35,7 @@
   * Removed reference to infinite list
   * Updated props table to better format types, defaults and description.
   * Added postcss as dev dependency.
+  * Update eslint and stylelint.
 
 ## 6.30.0 - (September 11, 2020)