+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Removed
-  * Removed terra-toolkit from peerDependencies
+  * Removed terra-toolkit from peer dependency
 
 * Changed
   * Updated to use terra-open-source-scripts to release the project