+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   * Removed terra-toolkit from peer dependency
 
 * Changed
+  * Updated wdio screenshots to fix main build
+
+* Changed
   * Updated to use terra-open-source-scripts to release the project
 
 ## 6.33.0 - (December 1, 2020)